+++ mlm_web_dioxus/src/components/mod.rs
@@ -1,9 +1,17 @@
 mod action_button;
 mod download_buttons;
+mod filter_controls;
 mod pagination;
+mod query_params;
+mod table_view;
 mod task_box;
 
 pub use action_button::ActionButton;
 pub use download_buttons::{DownloadButtonMode, DownloadButtons, SimpleDownloadButtons};
+pub use filter_controls::{
+    ActiveFilterChip, ActiveFilters, ColumnSelector, ColumnToggleOption, PageSizeSelector,
+};
 pub use pagination::Pagination;
+pub use query_params::{apply_click_filter, build_query_string, parse_location_query_pairs};
+pub use table_view::TableView;
 pub use task_box::TaskBox;
