+++ mlm_web_dioxus/src/components/table_view.rs
@@ -1,0 +1,8 @@
+use dioxus::prelude::*;
+
+#[component]
+pub fn TableView(class: String, style: String, children: Element) -> Element {
+    rsx! {
+        div { class: "{class}", style: "{style}", {children} }
+    }
+}
