+++ mlm_web_dioxus/src/torrents.rs
@@ -1,9 +1,4 @@
-use std::collections::BTreeSet;
-
-use crate::components::Pagination;
 use dioxus::prelude::*;
-#[cfg(feature = "web")]
-use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "server")]
@@ -800,1266 +795,7 @@ fn fuzzy_score(query: &str, target: &str) -> isize {
         .map_or(0, |m: sublime_fuzzy::Match| m.score())
 }
 
-fn filter_name(filter: TorrentsPageFilter) -> &'static str {
-    match filter {
-        TorrentsPageFilter::Kind => "Type",
-        TorrentsPageFilter::Category => "Category",
-        TorrentsPageFilter::Categories => "Categories",
-        TorrentsPageFilter::Flags => "Flags",
-        TorrentsPageFilter::Title => "Title",
-        TorrentsPageFilter::Author => "Authors",
-        TorrentsPageFilter::Narrator => "Narrators",
-        TorrentsPageFilter::Series => "Series",
-        TorrentsPageFilter::Language => "Language",
-        TorrentsPageFilter::Filetype => "Filetypes",
-        TorrentsPageFilter::Linker => "Linker",
-        TorrentsPageFilter::QbitCategory => "Qbit Category",
-        TorrentsPageFilter::Linked => "Linked",
-        TorrentsPageFilter::LibraryMismatch => "Library mismatch",
-        TorrentsPageFilter::ClientStatus => "Client status",
-        TorrentsPageFilter::Abs => "ABS",
-        TorrentsPageFilter::Query => "Query",
-        TorrentsPageFilter::Source => "Source",
-        TorrentsPageFilter::Metadata => "Metadata",
-    }
-}
+mod components;
+mod query;
 
-fn flag_icon(flag: &str) -> Option<(&'static str, &'static str)> {
-    match flag {
-        "language" => Some(("/assets/icons/language.png", "Crude Language")),
-        "violence" => Some(("/assets/icons/hand.png", "Violence")),
-        "some_explicit" => Some((
-            "/assets/icons/lipssmall.png",
-            "Some Sexually Explicit Content",
-        )),
-        "explicit" => Some(("/assets/icons/flames.png", "Sexually Explicit Content")),
-        "abridged" => Some(("/assets/icons/abridged.png", "Abridged")),
-        "lgbt" => Some(("/assets/icons/lgbt.png", "LGBT")),
-        _ => None,
-    }
-}
-
-fn apply_filter(
-    filters: &mut Signal<Vec<(TorrentsPageFilter, String)>>,
-    field: TorrentsPageFilter,
-    value: String,
-) {
-    let mut next = filters.read().clone();
-    next.retain(|(f, _)| *f != field);
-    next.push((field, value));
-    filters.set(next);
-}
-
-#[derive(Clone)]
-struct LegacyQueryState {
-    query: String,
-    sort: Option<TorrentsPageSort>,
-    asc: bool,
-    filters: Vec<(TorrentsPageFilter, String)>,
-    from: usize,
-    page_size: usize,
-    show: TorrentsPageColumns,
-}
-
-impl Default for LegacyQueryState {
-    fn default() -> Self {
-        Self {
-            query: String::new(),
-            sort: None,
-            asc: false,
-            filters: Vec::new(),
-            from: 0,
-            page_size: 500,
-            show: TorrentsPageColumns::default(),
-        }
-    }
-}
-
-#[cfg(feature = "web")]
-fn parse_query_enum<T: DeserializeOwned>(value: &str) -> Option<T> {
-    serde_json::from_str::<T>(&format!("\"{value}\"")).ok()
-}
-
-fn encode_query_enum<T: Serialize>(value: T) -> Option<String> {
-    serde_json::to_string(&value)
-        .ok()
-        .map(|raw| raw.trim_matches('"').to_string())
-}
-
-#[cfg(feature = "web")]
-fn decode_query_value(value: &str) -> String {
-    let replaced = value.replace('+', " ");
-    urlencoding::decode(&replaced)
-        .map(|s| s.to_string())
-        .unwrap_or(replaced)
-}
-
-fn show_to_query_value(show: TorrentsPageColumns) -> String {
-    let mut values = Vec::new();
-    if show.category {
-        values.push("category");
-    }
-    if show.categories {
-        values.push("categories");
-    }
-    if show.flags {
-        values.push("flags");
-    }
-    if show.edition {
-        values.push("edition");
-    }
-    if show.authors {
-        values.push("author");
-    }
-    if show.narrators {
-        values.push("narrator");
-    }
-    if show.series {
-        values.push("series");
-    }
-    if show.language {
-        values.push("language");
-    }
-    if show.size {
-        values.push("size");
-    }
-    if show.filetypes {
-        values.push("filetype");
-    }
-    if show.linker {
-        values.push("linker");
-    }
-    if show.qbit_category {
-        values.push("qbit_category");
-    }
-    if show.path {
-        values.push("path");
-    }
-    if show.created_at {
-        values.push("created_at");
-    }
-    if show.uploaded_at {
-        values.push("uploaded_at");
-    }
-    values.join(",")
-}
-
-#[cfg(feature = "web")]
-fn show_from_query_value(value: &str) -> TorrentsPageColumns {
-    let mut show = TorrentsPageColumns {
-        category: false,
-        categories: false,
-        flags: false,
-        edition: false,
-        authors: false,
-        narrators: false,
-        series: false,
-        language: false,
-        size: false,
-        filetypes: false,
-        linker: false,
-        qbit_category: false,
-        path: false,
-        created_at: false,
-        uploaded_at: false,
-    };
-    for item in value.split(',') {
-        match item {
-            "category" => show.category = true,
-            "categories" => show.categories = true,
-            "flags" => show.flags = true,
-            "edition" => show.edition = true,
-            "author" => show.authors = true,
-            "narrator" => show.narrators = true,
-            "series" => show.series = true,
-            "language" => show.language = true,
-            "size" => show.size = true,
-            "filetype" => show.filetypes = true,
-            "linker" => show.linker = true,
-            "qbit_category" => show.qbit_category = true,
-            "path" => show.path = true,
-            "created_at" => show.created_at = true,
-            "uploaded_at" => show.uploaded_at = true,
-            _ => {}
-        }
-    }
-    show
-}
-
-fn parse_legacy_query_state() -> LegacyQueryState {
-    #[cfg(feature = "web")]
-    {
-        let mut state = LegacyQueryState::default();
-        let Some(window) = web_sys::window() else {
-            return state;
-        };
-        let Ok(search) = window.location().search() else {
-            return state;
-        };
-        let search = search.trim_start_matches('?');
-        if search.is_empty() {
-            return state;
-        }
-        for pair in search.split('&') {
-            let (raw_key, raw_value) = pair.split_once('=').unwrap_or((pair, ""));
-            let key = decode_query_value(raw_key);
-            let value = decode_query_value(raw_value);
-            match key.as_str() {
-                "sort_by" => {
-                    state.sort = parse_query_enum::<TorrentsPageSort>(&value);
-                }
-                "asc" => {
-                    state.asc = value == "true";
-                }
-                "from" => {
-                    if let Ok(v) = value.parse::<usize>() {
-                        state.from = v;
-                    }
-                }
-                "page_size" => {
-                    if let Ok(v) = value.parse::<usize>() {
-                        state.page_size = v;
-                    }
-                }
-                "show" => {
-                    state.show = show_from_query_value(&value);
-                }
-                "query" => {
-                    state.query = value;
-                }
-                _ => {
-                    if let Some(field) = parse_query_enum::<TorrentsPageFilter>(&key) {
-                        state.filters.push((field, value));
-                    }
-                }
-            }
-        }
-        state
-    }
-    #[cfg(not(feature = "web"))]
-    {
-        LegacyQueryState::default()
-    }
-}
-
-fn build_legacy_query_string(
-    query: &str,
-    sort: Option<TorrentsPageSort>,
-    asc: bool,
-    filters: &[(TorrentsPageFilter, String)],
-    from: usize,
-    page_size: usize,
-    show: TorrentsPageColumns,
-) -> String {
-    let mut params: Vec<(String, String)> = Vec::new();
-    if let Some(sort) = sort.and_then(encode_query_enum) {
-        params.push(("sort_by".to_string(), sort));
-    }
-    if asc {
-        params.push(("asc".to_string(), "true".to_string()));
-    }
-    if from > 0 {
-        params.push(("from".to_string(), from.to_string()));
-    }
-    if page_size != 500 {
-        params.push(("page_size".to_string(), page_size.to_string()));
-    }
-    if show != TorrentsPageColumns::default() {
-        params.push(("show".to_string(), show_to_query_value(show)));
-    }
-    if !query.is_empty() {
-        params.push(("query".to_string(), query.to_string()));
-    }
-    for (field, value) in filters {
-        if let Some(name) = encode_query_enum(*field) {
-            params.push((name, value.clone()));
-        }
-    }
-    params
-        .into_iter()
-        .map(|(k, v)| format!("{}={}", urlencoding::encode(&k), urlencoding::encode(&v)))
-        .collect::<Vec<_>>()
-        .join("&")
-}
-
-#[component]
-pub fn TorrentsPage() -> Element {
-    let mut query_input = use_signal(String::new);
-    let mut submitted_query = use_signal(String::new);
-    let mut sort = use_signal(|| None::<TorrentsPageSort>);
-    let mut asc = use_signal(|| false);
-    let mut filters = use_signal(Vec::<(TorrentsPageFilter, String)>::new);
-    let mut from = use_signal(|| 0usize);
-    let mut page_size = use_signal(|| 500usize);
-    let mut show = use_signal(TorrentsPageColumns::default);
-    let mut selected = use_signal(BTreeSet::<String>::new);
-    let mut status_msg = use_signal(|| None::<(String, bool)>);
-    let mut cached = use_signal(|| None::<TorrentsData>);
-    let loading_action = use_signal(|| false);
-    let mut last_request_key = use_signal(String::new);
-    let mut url_init_done = use_signal(|| false);
-
-    let mut torrents_data = match use_server_future(move || async move {
-        let mut server_filters = filters.read().clone();
-        let query = submitted_query.read().trim().to_string();
-        if !query.is_empty() {
-            server_filters.push((TorrentsPageFilter::Query, query));
-        }
-        get_torrents_data(
-            *sort.read(),
-            *asc.read(),
-            server_filters,
-            Some(*from.read()),
-            Some(*page_size.read()),
-            *show.read(),
-        )
-        .await
-    }) {
-        Ok(resource) => resource,
-        Err(_) => {
-            return rsx! {
-                div { class: "torrents-page",
-                    div { class: "row",
-                        h1 { "Torrents" }
-                    }
-                    p { "Loading torrents..." }
-                }
-            };
-        }
-    };
-
-    let value = torrents_data.value();
-    let pending = torrents_data.pending();
-
-    {
-        let value = value.read();
-        if let Some(Ok(data)) = &*value {
-            cached.set(Some(data.clone()));
-        }
-    }
-
-    let data_to_show = {
-        let value = value.read();
-        match &*value {
-            Some(Ok(data)) => Some(data.clone()),
-            _ => cached.read().clone(),
-        }
-    };
-
-    use_effect(move || {
-        if *url_init_done.read() {
-            return;
-        }
-        let parsed = parse_legacy_query_state();
-        query_input.set(parsed.query.clone());
-        submitted_query.set(parsed.query);
-        sort.set(parsed.sort);
-        asc.set(parsed.asc);
-        filters.set(parsed.filters);
-        from.set(parsed.from);
-        page_size.set(parsed.page_size);
-        show.set(parsed.show);
-        url_init_done.set(true);
-    });
-
-    use_effect(move || {
-        if !*url_init_done.read() {
-            return;
-        }
-        let query = submitted_query.read().trim().to_string();
-        let sort = *sort.read();
-        let asc = *asc.read();
-        let filters = filters.read().clone();
-        let from = *from.read();
-        let page_size = *page_size.read();
-        let show = *show.read();
-
-        let query_string =
-            build_legacy_query_string(&query, sort, asc, &filters, from, page_size, show);
-        let should_restart = *last_request_key.read() != query_string;
-        if should_restart {
-            last_request_key.set(query_string.clone());
-            torrents_data.restart();
-        }
-    });
-
-    let sort_header = |label: &'static str, key: TorrentsPageSort| {
-        let active = *sort.read() == Some(key);
-        let arrow = if active {
-            if *asc.read() { "↑" } else { "↓" }
-        } else {
-            ""
-        };
-        rsx! {
-            div { class: "header",
-                button {
-                    r#type: "button",
-                    class: "link",
-                    onclick: {
-                        let mut sort = sort;
-                        let mut asc = asc;
-                        let mut from = from;
-                        move |_| {
-                            if *sort.read() == Some(key) {
-                                let next_asc = !*asc.read();
-                                asc.set(next_asc);
-                            } else {
-                                sort.set(Some(key));
-                                asc.set(false);
-                            }
-                            from.set(0);
-                        }
-                    },
-                    "{label}"
-                    "{arrow}"
-                }
-            }
-        }
-    };
-
-    rsx! {
-        div { class: "torrents-page",
-            form {
-                class: "row",
-                onsubmit: move |ev: Event<FormData>| {
-                    ev.prevent_default();
-                    submitted_query.set(query_input.read().trim().to_string());
-                    from.set(0);
-                },
-                h1 { "Torrents" }
-                label {
-                    input {
-                        r#type: "submit",
-                        value: "Search",
-                        style: "display: none;",
-                    }
-                    "Search: "
-                    input {
-                        r#type: "text",
-                        name: "query",
-                        value: "{query_input}",
-                        oninput: move |ev| query_input.set(ev.value()),
-                    }
-                    button {
-                        r#type: "button",
-                        onclick: move |_| {
-                            query_input.set(String::new());
-                            submitted_query.set(String::new());
-                            from.set(0);
-                        },
-                        "×"
-                    }
-                }
-                div { class: "table_options",
-                    div { class: "option_group query",
-                        "Columns:"
-                        div {
-                            label {
-                                "Category"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().category,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.category = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Categories"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().categories,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.categories = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Flags"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().flags,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.flags = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Edition"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().edition,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.edition = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Authors"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().authors,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.authors = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Narrators"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().narrators,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.narrators = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Series"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().series,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.series = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Language"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().language,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.language = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Size"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().size,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.size = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Filetypes"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().filetypes,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.filetypes = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Linker"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().linker,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.linker = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Qbit Category"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().qbit_category,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.qbit_category = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Path"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().path,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.path = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Added At"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().created_at,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.created_at = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                            label {
-                                "Uploaded At"
-                                input {
-                                    r#type: "checkbox",
-                                    checked: show.read().uploaded_at,
-                                    onchange: move |ev| {
-                                        let mut next = *show.read();
-                                        next.uploaded_at = ev.value() == "true";
-                                        show.set(next);
-                                    },
-                                }
-                            }
-                        }
-                    }
-                    div { class: "option_group query",
-                        "Page size: "
-                        select {
-                            value: "{page_size}",
-                            onchange: move |ev| {
-                                if let Ok(v) = ev.value().parse::<usize>() {
-                                    page_size.set(v);
-                                    from.set(0);
-                                }
-                            },
-                            option { value: "100", "100" }
-                            option { value: "500", "500" }
-                            option { value: "1000", "1000" }
-                            option { value: "5000", "5000" }
-                            option { value: "0", "all" }
-                        }
-                    }
-                }
-            }
-
-            if let Some((msg, is_error)) = status_msg.read().as_ref() {
-                p { class: if *is_error { "error" } else { "loading-indicator" },
-                    "{msg}"
-                    button {
-                        r#type: "button",
-                        style: "margin-left: 10px; cursor: pointer;",
-                        onclick: move |_| status_msg.set(None),
-                        "⨯"
-                    }
-                }
-            }
-
-            div { class: "option_group query",
-                if !submitted_query.read().is_empty() {
-                    span { class: "item",
-                        "Query: {submitted_query}"
-                        button {
-                            r#type: "button",
-                            onclick: move |_| {
-                                submitted_query.set(String::new());
-                                query_input.set(String::new());
-                                from.set(0);
-                            },
-                            " ×"
-                        }
-                    }
-                }
-                for (field , value) in filters.read().clone() {
-                    span { class: "item",
-                        "{filter_name(field)}: {value}"
-                        button {
-                            r#type: "button",
-                            onclick: {
-                                let value = value.clone();
-                                move |_| {
-                                    filters.write().retain(|(f, v)| !(*f == field && *v == value));
-                                    from.set(0);
-                                }
-                            },
-                            " ×"
-                        }
-                    }
-                }
-                if !filters.read().is_empty() || !submitted_query.read().is_empty() {
-                    button {
-                        r#type: "button",
-                        onclick: move |_| {
-                            filters.set(Vec::new());
-                            submitted_query.set(String::new());
-                            query_input.set(String::new());
-                            from.set(0);
-                        },
-                        "Clear filters"
-                    }
-                }
-            }
-
-            if let Some(data) = data_to_show {
-                if data.torrents.is_empty() {
-                    p {
-                        i { "You have no torrents selected by MLM" }
-                    }
-                } else {
-                    div { class: "actions actions_torrent",
-                        for action in [
-                            TorrentsBulkAction::Refresh,
-                            TorrentsBulkAction::RefreshRelink,
-                            TorrentsBulkAction::Clean,
-                            TorrentsBulkAction::Remove,
-                        ]
-                        {
-                            button {
-                                r#type: "button",
-                                disabled: *loading_action.read(),
-                                onclick: {
-                                    let mut loading_action = loading_action;
-                                    let mut status_msg = status_msg;
-                                    let mut torrents_data = torrents_data;
-                                    let mut selected = selected;
-                                    move |_| {
-                                        let ids: Vec<String> = selected.read().iter().cloned().collect();
-                                        if ids.is_empty() {
-                                            status_msg.set(Some(("Select at least one torrent".to_string(), true)));
-                                            return;
-                                        }
-                                        loading_action.set(true);
-                                        status_msg.set(None);
-                                        spawn(async move {
-                                            match apply_torrents_action(action, ids).await {
-                                                Ok(_) => {
-                                                    status_msg
-                                                        .set(Some((action.success_label().to_string(), false)));
-                                                    selected.set(BTreeSet::new());
-                                                    torrents_data.restart();
-                                                }
-                                                Err(e) => {
-                                                    status_msg
-                                                        .set(
-                                                            Some((format!("{} failed: {e}", action.label()), true)),
-                                                        );
-                                                }
-                                            }
-                                            loading_action.set(false);
-                                        });
-                                    }
-                                },
-                                "{action.label()}"
-                            }
-                        }
-                    }
-
-                    if pending && cached.read().is_some() {
-                        p { class: "loading-indicator", "Refreshing torrent list..." }
-                    }
-                    div {
-                        class: "TorrentsTable table2",
-                        style: "--torrents-grid: {show.read().table_grid_template()};",
-                        {
-                            let all_selected = data
-                                .torrents
-                                .iter()
-                                .all(|torrent| selected.read().contains(&torrent.id));
-                            rsx! {
-                                div { class: "torrents-grid-row",
-                                    div { class: "header",
-                                        input {
-                                            r#type: "checkbox",
-                                            checked: all_selected,
-                                            onchange: {
-                                                let row_ids = data
-                                                    .torrents
-                                                    .iter()
-                                                    .map(|torrent| torrent.id.clone())
-                                                    .collect::<Vec<_>>();
-                                                move |ev| {
-                                                    if ev.value() == "true" {
-                                                        let mut next = selected.read().clone();
-                                                        for id in &row_ids {
-                                                            next.insert(id.clone());
-                                                        }
-                                                        selected.set(next);
-                                                    } else {
-                                                        let mut next = selected.read().clone();
-                                                        for id in &row_ids {
-                                                            next.remove(id);
-                                                        }
-                                                        selected.set(next);
-                                                    }
-                                                }
-                                            },
-                                        }
-                                    }
-                                    {sort_header("Type", TorrentsPageSort::Kind)}
-                                    if show.read().categories {
-                                        div { class: "header", "Categories" }
-                                    }
-                                    if show.read().flags {
-                                        div { class: "header", "Flags" }
-                                    }
-                                    {sort_header("Title", TorrentsPageSort::Title)}
-                                    if show.read().edition {
-                                        {sort_header("Edition", TorrentsPageSort::Edition)}
-                                    }
-                                    if show.read().authors {
-                                        {sort_header("Authors", TorrentsPageSort::Authors)}
-                                    }
-                                    if show.read().narrators {
-                                        {sort_header("Narrators", TorrentsPageSort::Narrators)}
-                                    }
-                                    if show.read().series {
-                                        {sort_header("Series", TorrentsPageSort::Series)}
-                                    }
-                                    if show.read().language {
-                                        {sort_header("Language", TorrentsPageSort::Language)}
-                                    }
-                                    if show.read().size {
-                                        {sort_header("Size", TorrentsPageSort::Size)}
-                                    }
-                                    if show.read().filetypes {
-                                        div { class: "header", "Filetypes" }
-                                    }
-                                    if show.read().linker {
-                                        {sort_header("Linker", TorrentsPageSort::Linker)}
-                                    }
-                                    if show.read().qbit_category {
-                                        {sort_header("Qbit Category", TorrentsPageSort::QbitCategory)}
-                                    }
-                                    {
-                                        sort_header(
-                                            if show.read().path { "Path" } else { "Linked" },
-                                            TorrentsPageSort::Linked,
-                                        )
-                                    }
-                                    if show.read().created_at {
-                                        {sort_header("Added At", TorrentsPageSort::CreatedAt)}
-                                    }
-                                    if show.read().uploaded_at {
-                                        {sort_header("Uploaded At", TorrentsPageSort::UploadedAt)}
-                                    }
-                                    div { class: "header", "" }
-                                }
-                            }
-                        }
-
-                        for torrent in data.torrents.clone() {
-                            {
-                                let row_id = torrent.id.clone();
-                                let row_selected = selected.read().contains(&row_id);
-                                rsx! {
-                                    div { class: "torrents-grid-row", key: "{row_id}",
-                                        div {
-                                            input {
-                                                r#type: "checkbox",
-                                                checked: row_selected,
-                                                onchange: {
-                                                    let row_id = row_id.clone();
-                                                    move |ev| {
-                                                        let mut next = selected.read().clone();
-                                                        if ev.value() == "true" {
-                                                            next.insert(row_id.clone());
-                                                        } else {
-                                                            next.remove(&row_id);
-                                                        }
-                                                        selected.set(next);
-                                                    }
-                                                },
-                                            }
-                                        }
-                                        div {
-                                            button {
-                                                r#type: "button",
-                                                class: "link",
-                                                title: "{torrent.meta.cat_name}",
-                                                onclick: {
-                                                    let value = torrent.meta.media_type.clone();
-                                                    move |_| {
-                                                        apply_filter(&mut filters, TorrentsPageFilter::Kind, value.clone());
-                                                        from.set(0);
-                                                    }
-                                                },
-                                                "{torrent.meta.media_type}"
-                                            }
-                                            if show.read().category {
-                                                if let Some(cat_id) = torrent.meta.cat_id.clone() {
-                                                    div {
-                                                        button {
-                                                            r#type: "button",
-                                                            class: "link",
-                                                            onclick: {
-                                                                let label = cat_id.clone();
-                                                                move |_| {
-                                                                    apply_filter(&mut filters, TorrentsPageFilter::Category, label.clone());
-                                                                    from.set(0);
-                                                                }
-                                                            },
-                                                            "{torrent.meta.cat_name}"
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().categories {
-                                            div {
-                                                for category in torrent.meta.categories.clone() {
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: {
-                                                            let category = category.clone();
-                                                            move |_| {
-                                                                apply_filter(&mut filters, TorrentsPageFilter::Categories, category.clone());
-                                                                from.set(0);
-                                                            }
-                                                        },
-                                                        "{category}"
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().flags {
-                                            div {
-                                                for flag in torrent.meta.flags.clone() {
-                                                    if let Some((src, title)) = flag_icon(&flag) {
-                                                        button {
-                                                            r#type: "button",
-                                                            class: "link",
-                                                            onclick: {
-                                                                let flag = flag.clone();
-                                                                move |_| {
-                                                                    apply_filter(&mut filters, TorrentsPageFilter::Flags, flag.clone());
-                                                                    from.set(0);
-                                                                }
-                                                            },
-                                                            img {
-                                                                class: "flag",
-                                                                src: "{src}",
-                                                                alt: "{title}",
-                                                                title: "{title}",
-                                                            }
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        div {
-                                            button {
-                                                r#type: "button",
-                                                class: "link",
-                                                onclick: {
-                                                    let title = torrent.meta.title.clone();
-                                                    move |_| {
-                                                        apply_filter(&mut filters, TorrentsPageFilter::Title, title.clone());
-                                                        from.set(0);
-                                                    }
-                                                },
-                                                "{torrent.meta.title}"
-                                            }
-                                            if torrent.client_status.as_deref() == Some("removed_from_tracker") {
-                                                span {
-                                                    class: "warn",
-                                                    title: "Torrent is removed from tracker but still seeding",
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: move |_| {
-                                                            apply_filter(
-                                                                &mut filters,
-                                                                TorrentsPageFilter::ClientStatus,
-                                                                "removed_from_tracker".to_string(),
-                                                            );
-                                                            from.set(0);
-                                                        },
-                                                        "⚠"
-                                                    }
-                                                }
-                                            }
-                                            if torrent.client_status.as_deref() == Some("not_in_client") {
-                                                span { title: "Torrent is not seeding",
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: move |_| {
-                                                            apply_filter(
-                                                                &mut filters,
-                                                                TorrentsPageFilter::ClientStatus,
-                                                                "not_in_client".to_string(),
-                                                            );
-                                                            from.set(0);
-                                                        },
-                                                        "ℹ"
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().edition {
-                                            div { "{torrent.meta.edition.clone().unwrap_or_default()}" }
-                                        }
-                                        if show.read().authors {
-                                            div {
-                                                for author in torrent.meta.authors.clone() {
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: {
-                                                            let author = author.clone();
-                                                            move |_| {
-                                                                apply_filter(&mut filters, TorrentsPageFilter::Author, author.clone());
-                                                                from.set(0);
-                                                            }
-                                                        },
-                                                        "{author}"
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().narrators {
-                                            div {
-                                                for narrator in torrent.meta.narrators.clone() {
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: {
-                                                            let narrator = narrator.clone();
-                                                            move |_| {
-                                                                apply_filter(&mut filters, TorrentsPageFilter::Narrator, narrator.clone());
-                                                                from.set(0);
-                                                            }
-                                                        },
-                                                        "{narrator}"
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().series {
-                                            div {
-                                                for series in torrent.meta.series.clone() {
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: {
-                                                            let series_name = series.name.clone();
-                                                            move |_| {
-                                                                apply_filter(&mut filters, TorrentsPageFilter::Series, series_name.clone());
-                                                                from.set(0);
-                                                            }
-                                                        },
-                                                        if series.entries.is_empty() {
-                                                            "{series.name}"
-                                                        } else {
-                                                            "{series.name} #{series.entries}"
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().language {
-                                            div {
-                                                button {
-                                                    r#type: "button",
-                                                    class: "link",
-                                                    onclick: {
-                                                        let value = torrent.meta.language.clone().unwrap_or_default();
-                                                        move |_| {
-                                                            apply_filter(&mut filters, TorrentsPageFilter::Language, value.clone());
-                                                            from.set(0);
-                                                        }
-                                                    },
-                                                    "{torrent.meta.language.clone().unwrap_or_default()}"
-                                                }
-                                            }
-                                        }
-                                        if show.read().size {
-                                            div { "{torrent.meta.size}" }
-                                        }
-                                        if show.read().filetypes {
-                                            div {
-                                                for filetype in torrent.meta.filetypes.clone() {
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: {
-                                                            let filetype = filetype.clone();
-                                                            move |_| {
-                                                                apply_filter(&mut filters, TorrentsPageFilter::Filetype, filetype.clone());
-                                                                from.set(0);
-                                                            }
-                                                        },
-                                                        "{filetype}"
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().linker {
-                                            div {
-                                                button {
-                                                    r#type: "button",
-                                                    class: "link",
-                                                    onclick: {
-                                                        let linker = torrent.linker.clone().unwrap_or_default();
-                                                        move |_| {
-                                                            apply_filter(&mut filters, TorrentsPageFilter::Linker, linker.clone());
-                                                            from.set(0);
-                                                        }
-                                                    },
-                                                    "{torrent.linker.clone().unwrap_or_default()}"
-                                                }
-                                            }
-                                        }
-                                        if show.read().qbit_category {
-                                            div {
-                                                button {
-                                                    r#type: "button",
-                                                    class: "link",
-                                                    onclick: {
-                                                        let category = torrent.category.clone().unwrap_or_default();
-                                                        move |_| {
-                                                            apply_filter(
-                                                                &mut filters,
-                                                                TorrentsPageFilter::QbitCategory,
-                                                                category.clone(),
-                                                            );
-                                                            from.set(0);
-                                                        }
-                                                    },
-                                                    "{torrent.category.clone().unwrap_or_default()}"
-                                                }
-                                            }
-                                        }
-                                        if show.read().path {
-                                            div {
-                                                "{torrent.library_path.clone().unwrap_or_default()}"
-                                                if let Some(mismatch) = torrent.library_mismatch.clone() {
-                                                    span { class: "warn", title: "{mismatch.title()}",
-                                                        button {
-                                                            r#type: "button",
-                                                            class: "link",
-                                                            onclick: move |_| {
-                                                                apply_filter(
-                                                                    &mut filters,
-                                                                    TorrentsPageFilter::LibraryMismatch,
-                                                                    mismatch.filter_value().to_string(),
-                                                                );
-                                                                from.set(0);
-                                                            },
-                                                            "⚠"
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        } else {
-                                            div {
-                                                if let Some(path) = torrent.library_path.clone() {
-                                                    span { title: "{path}",
-                                                        button {
-                                                            r#type: "button",
-                                                            class: "link",
-                                                            onclick: {
-                                                                let linked = torrent.linked;
-                                                                move |_| {
-                                                                    apply_filter(&mut filters, TorrentsPageFilter::Linked, linked.to_string());
-                                                                    from.set(0);
-                                                                }
-                                                            },
-                                                            "{torrent.linked}"
-                                                        }
-                                                    }
-                                                } else {
-                                                    button {
-                                                        r#type: "button",
-                                                        class: "link",
-                                                        onclick: {
-                                                            let linked = torrent.linked;
-                                                            move |_| {
-                                                                apply_filter(&mut filters, TorrentsPageFilter::Linked, linked.to_string());
-                                                                from.set(0);
-                                                            }
-                                                        },
-                                                        "{torrent.linked}"
-                                                    }
-                                                }
-                                                if let Some(mismatch) = torrent.library_mismatch.clone() {
-                                                    span { class: "warn", title: "{mismatch.title()}",
-                                                        button {
-                                                            r#type: "button",
-                                                            class: "link",
-                                                            onclick: move |_| {
-                                                                apply_filter(
-                                                                    &mut filters,
-                                                                    TorrentsPageFilter::LibraryMismatch,
-                                                                    mismatch.filter_value().to_string(),
-                                                                );
-                                                                from.set(0);
-                                                            },
-                                                            "⚠"
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if show.read().created_at {
-                                            div { "{torrent.created_at}" }
-                                        }
-                                        if show.read().uploaded_at {
-                                            div { "{torrent.uploaded_at}" }
-                                        }
-                                        div {
-                                            a { href: "/dioxus/torrents/{torrent.id}", "open" }
-                                            if let Some(mam_id) = torrent.mam_id {
-                                                a {
-                                                    href: "https://www.myanonamouse.net/t/{mam_id}",
-                                                    target: "_blank",
-                                                    "MaM"
-                                                }
-                                            }
-                                            if let (Some(abs_url), Some(abs_id)) = (&data.abs_url, &torrent.abs_id) {
-                                                a {
-                                                    href: "{abs_url}/audiobookshelf/item/{abs_id}",
-                                                    target: "_blank",
-                                                    "ABS"
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    p { class: "faint",
-                        "Showing {data.from} to {data.from + data.torrents.len()} of {data.total}"
-                    }
-                    Pagination {
-                        total: data.total,
-                        from: data.from,
-                        page_size: data.page_size,
-                        on_change: move |new_from| {
-                            from.set(new_from);
-                        },
-                    }
-                }
-            } else if let Some(Err(e)) = &*value.read() {
-                p { class: "error", "Error: {e}" }
-            } else {
-                p { "Loading torrents..." }
-            }
-        }
-    }
-}
+pub use components::TorrentsPage;
